+++ solutions/alt-tag-generator/pages/api/generate.ts
@@ -1,53 +1,25 @@
-import type { NextApiRequest, NextApiResponse } from "next";
+import { NextRequest, NextResponse } from 'next/server'
+import Replicate from 'replicate'
 
-type Data = string;
-interface ExtendedNextApiRequest extends NextApiRequest {
-  body: {
-    imageUrl: string;
-  };
+export const config = {
+  runtime: 'edge',
 }
 
-export default async function handler(
-  req: ExtendedNextApiRequest,
-  res: NextApiResponse<Data>
-) {
-  const { imageUrl } = req.query;
-  // POST request to Replicate to start the alt generation process
-  let startResponse = await fetch("https://api.replicate.com/v1/predictions", {
-    method: "POST",
-    headers: {
-      "Content-Type": "application/json",
-      Authorization: "Token " + process.env.REPLICATE_API_KEY,
-    },
-    body: JSON.stringify({
-      version:
-        "2e1dddc8621f72155f24cf2e0adbde548458d3cab9f00c0139eea840d0ac4746",
-      input: { image: imageUrl },
-    }),
-  });
+const replicate = new Replicate({
+  auth: process.env.REPLICATE_API_TOKEN || '',
+})
 
-  let jsonStartResponse = await startResponse.json();
-  let endpointUrl = jsonStartResponse.urls.get;
+export default async function handler(req: NextRequest) {
+  const image =
+    req.nextUrl.searchParams.get('imageUrl') || 'https://dub.sh/confpic'
 
-  // GET request to get the status of the alt generation process & return the result when it's ready
-  let altText: string | null = null;
-  while (!altText) {
-    // Loop in 500ms intervals until the alt text is ready
-    let finalResponse = await fetch(endpointUrl, {
-      method: "GET",
-      headers: {
-        "Content-Type": "application/json",
-        Authorization: "Token " + process.env.REPLICATE_API_KEY,
+  const output = await replicate.run(
+    'salesforce/blip:2e1dddc8621f72155f24cf2e0adbde548458d3cab9f00c0139eea840d0ac4746',
+    {
+      input: {
+        image,
       },
-    });
-    let jsonFinalResponse = await finalResponse.json();
-    if (jsonFinalResponse.status === "succeeded") {
-      altText = jsonFinalResponse.output.split("Caption: ")[1];
-    } else if (jsonFinalResponse.status === "failed") {
-      break;
-    } else {
-      await new Promise((resolve) => setTimeout(resolve, 500));
     }
-  }
-  res.status(200).json(altText ? altText : "Failed to generate alt text");
+  )
+  return NextResponse.json(output)
 }