+++ storage/blob-starter/next.config.js
@@ -1,0 +1,8 @@
+/** @type {import('next').NextConfig} */
+const nextConfig = {
+  images: {
+    domains: ['public.blob.vercel-storage.com'],
+  },
+}
+
+module.exports = nextConfig