+++ storage/kv-redis-nuxt/nuxt.config.ts
@@ -1,0 +1,15 @@
+// https://nuxt.com/docs/api/configuration/nuxt-config
+export default defineNuxtConfig({
+  nitro: {
+    storage: {
+      data: { driver: 'vercelKV' },
+    },
+  },
+  css: ['~/assets/css/main.css'],
+  postcss: {
+    plugins: {
+      tailwindcss: {},
+      autoprefixer: {},
+    },
+  },
+})