+++ storage/postgres-drizzle/next.config.js
@@ -1,0 +1,8 @@
+/** @type {import('next').NextConfig} */
+const nextConfig = {
+  images: {
+    domains: ['images.ctfassets.net'],
+  },
+}
+
+module.exports = nextConfig